+++ src/Utility/math.cpp
@@ -2,9 +2,9 @@
 
 glm::vec3 clickray(double mouse_x, double mouse_y, double width, double height, const glm::mat4& projection, Camera& camera)
 {
-	float x = (2.0f * mouse_x) / width - 1.0f;
-	float y = 1.0f - (2.0f * mouse_y) / height;
-	float z = 1.0f;
+	double x = (2.0f * mouse_x) / width - 1.0f;
+	double y = 1.0f - (2.0f * mouse_y) / height;
+	double z = 1.0f;
 	//dbgprint("screen coords: %f %f\n", x, y);
 	glm::vec3 ray_nds = glm::vec3(x, y, z);
 	glm::vec4 ray_clip = glm::vec4(ray_nds.x, ray_nds.y, -1.0f, 1.0f);
@@ -17,9 +17,9 @@ glm::vec3 clickray(double mouse_x, double mouse_y, double width, double height, 
 
 bool ray_intersects_triangle(const glm::vec3& ray_origin, const glm::vec3& ray_vector, const glm::vec3& vertex0, const glm::vec3& vertex1, const glm::vec3& vertex2, glm::vec3& outIntersectionPoint)
 {
-	const float EPSILON = 0.0000001;
+	const double EPSILON = 0.0000001;
 	glm::vec3 edge1, edge2, h, s, q;
-	float a, f, u, v;
+	double a, f, u, v;
 	edge1 = vertex1 - vertex0;
 	edge2 = vertex2 - vertex0;
 	h = glm::cross(ray_vector, edge2);
@@ -36,10 +36,10 @@ bool ray_intersects_triangle(const glm::vec3& ray_origin, const glm::vec3& ray_v
 	if (v < 0.0 || u + v > 1.0)
 		return false;
 	// At this stage we can compute t to find out where the intersection point is on the line.
-	float t = f * glm::dot(edge2, q);
+	double t = f * glm::dot(edge2, q);
 	if (t > EPSILON) // ray intersection
 	{
-		outIntersectionPoint = ray_origin + ray_vector * t;
+		outIntersectionPoint = ray_origin + ray_vector * static_cast<float>(t);
 		return true;
 	}
 	else // This means that there is a line intersection but not a ray intersection.
